Added: Verwaltung der GpxFilenames und Categories (wie in WinCB) in der Datenbank Caches die durch den API-Import hinzugefügt werden werden in dem/der GpxFilename/Category "API-Import" zugeordnet
Added: Management of the GpxFilenames and Categories (like in WinCB) in the Database
Caches added by the API-Import will be added to the GpxFilename/Category "API-Import"

git-svn-id: svn://svn.code.sf.net/p/droidcachebox/code/trunk@338 5890f383-b09f-485f-8f59-22bc967d67fd

+++ Android_GUI/src/de/droidcachebox/DAO/CacheDAO.java
@@ -65,7 +65,7 @@ public class CacheDAO {
         	cache.TourName = "";
 
         if (reader.getString(21) != "")
-        	cache.GPXFilename_ID = reader.getInt(21);
+        	cache.GPXFilename_ID = reader.getLong(21);
         else
         	cache.GPXFilename_ID = -1;
